changed/added scripts - updated CMakeLists.txt LIBS (based on OS) - added gitignore

+++ build.sh
@@ -1,19 +1,22 @@
 #!/bin/bash
 
-cd "$(dirname "$0")"
+cd "$(dirname "$0")" || exit
 
-rm -r build/
-rm -r bin/
-mkdir build
-mkdir bin
+if [[ ! -d build/ ]]; then
+    mkdir build/
+fi
+if [[ -d bin/ ]]; then
+    rm -r bin/
+fi
+mkdir bin/
 
-cd build/
+cd build/ || exit
 
-cmake ../
+#TODO: change release type
+cmake -DCMAKE_BUILD_TYPE=Release ../
+#TODO: change make thread count
 make -j4
 
 cd ../
 #TODO: replace name of project
 cp build/<binary_name> bin/
-
-rm -r build/
